refactor: Refactor implementations and clean up styles
